pyproject.toml
[tool.poetry]
name = "ohlordy"
version = "0.1.0"
description = ""
authors = ["SimonStJG <Simon.StJG@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-vlc = "^3.0.7110"
"RPi.GPIO" = {version = "^0.7.0", extras = ["rpi"]}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"

[tool.poetry.scripts]
ohlordy = 'ohlordy:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
